+++ CHANGELOG.md
@@ -4,19 +4,29 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## Unreleased 2025-01-31
+
+### Changed
+
+- Replaced "overwrite existing graph" bool parameter with parameter with options replace, add, or stop if exists
 
 ## [1.0.0beta4] 2025-01-06
+
+### Fixed
 
 - fix errors for certain namespace prefixes 
 
 ## [1.0.0beta3] 2025-01-06
 
+### Changed
+
 - preserve multiple prefixes for one namespace in prefix.cc list
 
 ### [1.0.0beta2] 2024-12-03
 
+### Changed
 - removed dependency str2bool
-- use GraphParameterType for Shapes graph 
+  - use GraphParameterType for Shapes graph 
 
 
 ## [1.0.0beta1] 2024-07-24