+++ CHANGELOG.md
@@ -4,22 +4,31 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
-## [Unreleased]
+## [1.0.0] 2025-02-03
 
 ### Fixed
+
 - Fixed summary report.
 
 ### Added
+
 - Added shui:showAlways to all property shapes.
 - Introduced the `ignore properties` parameter to exclude specific properties from the shape graph.
+
+### Changed
+
+- documentation
+
 
 ## [1.0.0beta4] 2025-01-06
 
 - fix errors for certain namespace prefixes 
 
+
 ## [1.0.0beta3] 2025-01-06
 
 - preserve multiple prefixes for one namespace in prefix.cc list
+
 
 ### [1.0.0beta2] 2024-12-03